requirements.txt
scikit-learn==1.0.1
pandas
numpy
spacy
python-dotenv